requirements.txt
fastapi
uvicorn
python-dotenv
sqlalchemy
psycopg2-binary
passlib[bcrypt]
python-jose[cryptography]
alembic
pydantic-settings
pydantic[email]
google-genai
langchain_google_genai
langgraph
langchain
langmem
python-multipart
sqlalchemy-utils